fixup! refactor: meta 스키마 공유
